+++ part.go
@@ -7,11 +7,12 @@ import (
 	"fmt"
 	"io"
 	"io/ioutil"
-	"mime"
 	"mime/quotedprintable"
 	"net/textproto"
 	"strconv"
 	"strings"
+
+	"github.com/jhillyerd/enmime/internal/coding"
 )
 
 // Part represents a node in the MIME multipart tree.  The Content-Type, Disposition and File Name
@@ -98,6 +99,34 @@ func (p *Part) TextContent() bool {
 		strings.HasPrefix(p.ContentType, ctMultipartPrefix)
 }
 
+// setupHeaders reads the header, then populates the MIME header values for this Part.
+func (p *Part) setupHeaders(r *bufio.Reader, defaultContentType string) error {
+	header, err := readHeader(r, p)
+	if err != nil {
+		return err
+	}
+	p.Header = header
+	ctype := header.Get(hnContentType)
+	if ctype == "" {
+		if defaultContentType == "" {
+			p.addWarning(ErrorMissingContentType, "MIME parts should have a Content-Type header")
+			return nil
+		}
+		ctype = defaultContentType
+	}
+	// Parse Content-Type header
+	mtype, mparams, err := parseMediaType(ctype)
+	if err != nil {
+		return err
+	}
+	p.ContentType = mtype
+	// Set disposition, filename, charset if available
+	p.setupContentHeaders(mparams)
+	p.Boundary = mparams[hpBoundary]
+	p.ContentID = coding.FromIDHeader(header.Get(hnContentID))
+	return nil
+}
+
 // setupContentHeaders uses Content-Type media params and Content-Disposition headers to populate
 // the disposition, filename, and charset fields.
 func (p *Part) setupContentHeaders(mediaParams map[string]string) {
@@ -137,16 +166,16 @@ func (p *Part) buildContentReaders(r io.Reader) error {
 	p.rawReader = contentReader
 
 	// Allow later access to Base64 errors
-	var b64cleaner *base64Cleaner
+	var b64cleaner *coding.Base64Cleaner
 
 	// Build content decoding reader
 	encoding := p.Header.Get(hnContentEncoding)
 	switch strings.ToLower(encoding) {
 	case cteQuotedPrintable:
-		contentReader = newQPCleaner(contentReader)
+		contentReader = coding.NewQPCleaner(contentReader)
 		contentReader = quotedprintable.NewReader(contentReader)
 	case cteBase64:
-		b64cleaner = newBase64Cleaner(contentReader)
+		b64cleaner = coding.NewBase64Cleaner(contentReader)
 		contentReader = base64.NewDecoder(base64.RawStdEncoding, b64cleaner)
 	case cte8Bit, cte7Bit, cteBinary, "":
 		// No decoding required
@@ -163,7 +192,7 @@ func (p *Part) buildContentReaders(r io.Reader) error {
 	if valid && !detectAttachmentHeader(p.Header) {
 		// decodedReader is good; build character set conversion reader
 		if p.Charset != "" {
-			if reader, err := newCharsetReader(p.Charset, contentReader); err == nil {
+			if reader, err := coding.NewCharsetReader(p.Charset, contentReader); err == nil {
 				contentReader = reader
 			} else {
 				// Try to parse charset again here to see if we can salvage some badly formed ones
@@ -171,7 +200,7 @@ func (p *Part) buildContentReaders(r io.Reader) error {
 				charsetp := strings.Split(p.Charset, "=")
 				if strings.ToLower(charsetp[0]) == "charset" && len(charsetp) > 1 {
 					p.Charset = charsetp[1]
-					if reader, err := newCharsetReader(p.Charset, contentReader); err == nil {
+					if reader, err := coding.NewCharsetReader(p.Charset, contentReader); err == nil {
 						contentReader = reader
 					} else {
 						// Failed to get a conversion reader
@@ -186,10 +215,16 @@ func (p *Part) buildContentReaders(r io.Reader) error {
 	}
 	// Messy until Utf8Reader is removed
 	content, err := ioutil.ReadAll(contentReader)
-	p.Utf8Reader = contentReader
+	p.Utf8Reader = bytes.NewReader(content)
 	p.Content = content
 	if b64cleaner != nil {
-		p.Errors = append(p.Errors, b64cleaner.Errors...)
+		for _, err := range b64cleaner.Errors {
+			p.Errors = append(p.Errors, Error{
+				Name:   ErrorMalformedBase64,
+				Detail: err.Error(),
+				Severe: false,
+			})
+		}
 	}
 	return err
 }
@@ -197,103 +232,33 @@ func (p *Part) buildContentReaders(r io.Reader) error {
 // ReadParts reads a MIME document from the provided reader and parses it into tree of Part objects.
 func ReadParts(r io.Reader) (*Part, error) {
 	br := bufio.NewReader(r)
-	root := &Part{}
-
-	// Read header
-	header, err := readHeader(br, root)
+	root := &Part{PartID: "0"}
+	// Read header; top-level default CT is text/plain us-ascii according to RFC 822.
+	err := root.setupHeaders(br, `text/plain; charset="us-ascii"`)
 	if err != nil {
 		return nil, err
 	}
-	root.Header = header
-
-	// Content-Type, default is text/plain us-ascii according to RFC 822
-	mediatype := ctTextPlain
-	params := map[string]string{
-		"charset": "us-ascii",
-	}
-	contentType := header.Get(hnContentType)
-	if contentType != "" {
-		mediatype, params, err = parseMediaType(contentType)
-		if err != nil {
-			return nil, err
-		}
-	}
-	root.ContentType = mediatype
-	root.Charset = params[hpCharset]
-
-	if strings.HasPrefix(mediatype, ctMultipartPrefix) {
-		// Content is multipart, parse it
-		boundary := params[hpBoundary]
-		root.Boundary = boundary
+	if strings.HasPrefix(root.ContentType, ctMultipartPrefix) {
+		// Content is multipart, parse it.
 		err = parseParts(root, br)
 		if err != nil {
 			return nil, err
 		}
 	} else {
-		// Content is text or data, build content reader pipeline
+		// Content is text or data, build content reader pipeline.
 		if err := root.buildContentReaders(br); err != nil {
 			return nil, err
 		}
 	}
-
 	return root, nil
 }
 
-func parseMediaType(ctype string) (string, map[string]string, error) {
-	// Parse Content-Type header
-	mtype, mparams, err := mime.ParseMediaType(ctype)
-	if err != nil {
-		// Small hack to remove harmless charset duplicate params
-		mctype := parseBadContentType(ctype, ";")
-		mtype, mparams, err = mime.ParseMediaType(mctype)
-		if err != nil {
-			// Some badly formed content-types forget to send a ; between fields
-			mctype := parseBadContentType(ctype, " ")
-			if strings.Contains(mctype, `name=""`) {
-				mctype = strings.Replace(mctype, `name=""`, `name=" "`, -1)
-			}
-			mtype, mparams, err = mime.ParseMediaType(mctype)
-			if err != nil {
-				return "", make(map[string]string), err
-			}
-		}
-	}
-	return mtype, mparams, err
-}
-
-func parseBadContentType(ctype, sep string) string {
-	cp := strings.Split(ctype, sep)
-	mctype := ""
-	for _, p := range cp {
-		if strings.Contains(p, "=") {
-			params := strings.Split(p, "=")
-			if !strings.Contains(mctype, params[0]+"=") {
-				mctype += p + ";"
-			}
-		} else {
-			mctype += p + ";"
-		}
-	}
-	return mctype
-}
-
-// parseParts recursively parses a mime multipart document and sets each Part's PartID.
+// parseParts recursively parses a MIME multipart document and sets each Parts PartID.
 func parseParts(parent *Part, reader *bufio.Reader) error {
-	var prevSibling *Part
-
 	firstRecursion := parent.Parent == nil
-	// Set root PartID
-	if firstRecursion {
-		parent.PartID = "0"
-	}
-
-	var indexPartID int
-
-	// Loop over MIME parts
+	// Loop over MIME boundaries.
 	br := newBoundaryReader(reader, parent.Boundary)
-	for {
-		indexPartID++
-
+	for indexPartID := 1; true; indexPartID++ {
 		next, err := br.Next()
 		if err != nil && err != io.EOF {
 			return err
@@ -301,93 +266,57 @@ func parseParts(parent *Part, reader *bufio.Reader) error {
 		if !next {
 			break
 		}
-		p := &Part{Parent: parent}
-
-		// Set this Part's PartID, indicating its position within the MIME Part Tree
+		p := &Part{}
+		// Set this Part's PartID, indicating its position within the MIME Part tree.
 		if firstRecursion {
 			p.PartID = strconv.Itoa(indexPartID)
 		} else {
-			p.PartID = p.Parent.PartID + "." + strconv.Itoa(indexPartID)
+			p.PartID = parent.PartID + "." + strconv.Itoa(indexPartID)
 		}
-
+		// Look for part header.
 		bbr := bufio.NewReader(br)
-		header, err := readHeader(bbr, p)
-		p.Header = header
+		err = p.setupHeaders(bbr, "")
 		if err == errEmptyHeaderBlock {
 			// Empty header probably means the part didn't use the correct trailing "--" syntax to
 			// close its boundary.
 			if _, err = br.Next(); err != nil {
 				if err == io.EOF || strings.HasSuffix(err.Error(), "EOF") {
-					// There are no more Parts, but the error belongs to a sibling or parent,
-					// because this Part doesn't actually exist.
-					owner := parent
-					if prevSibling != nil {
-						owner = prevSibling
-					}
-					owner.addWarning(
-						ErrorMissingBoundary,
-						"Boundary %q was not closed correctly",
+					// There are no more Parts. The error must belong to the parent, because this
+					// part doesn't exist.
+					parent.addWarning(ErrorMissingBoundary, "Boundary %q was not closed correctly",
 						parent.Boundary)
 					break
 				}
-				return fmt.Errorf("Error at boundary %v: %v", parent.Boundary, err)
+				return fmt.Errorf("error at boundary %v: %v", parent.Boundary, err)
 			}
 		} else if err != nil {
 			return err
 		}
-
-		ctype := header.Get(hnContentType)
-		if ctype == "" {
-			p.addWarning(
-				ErrorMissingContentType,
-				"MIME parts should have a Content-Type header")
-		} else {
-			// Parse Content-Type header
-			mtype, mparams, err := parseMediaType(ctype)
-			if err != nil {
+		// Insert this Part into the MIME tree.
+		parent.AddChild(p)
+		if p.Boundary == "" {
+			// Content is text or data; build content reader pipeline.
+			if err := p.buildContentReaders(bbr); err != nil {
 				return err
 			}
-			p.ContentType = mtype
-
-			// Set disposition, filename, charset if available
-			p.setupContentHeaders(mparams)
-			p.Boundary = mparams[hpBoundary]
-		}
-
-		// Insert this Part into the MIME tree
-		if prevSibling != nil {
-			prevSibling.NextSibling = p
 		} else {
-			parent.FirstChild = p
-		}
-		prevSibling = p
-
-		if p.Boundary != "" {
-			// Content is another multipart
+			// Content is another multipart.
 			err = parseParts(p, bbr)
 			if err != nil {
 				return err
 			}
-		} else {
-			// Content is text or data: build content reader pipeline
-			if err := p.buildContentReaders(bbr); err != nil {
-				return err
-			}
 		}
 	}
-
-	// Store any content following the closing boundary marker into the epilogue
-	epilogue := new(bytes.Buffer)
-	if _, err := io.Copy(epilogue, reader); err != nil {
+	// Store any content following the closing boundary marker into the epilogue.
+	epilogue, err := ioutil.ReadAll(reader)
+	if err != nil {
 		return err
 	}
-	parent.Epilogue = epilogue.Bytes()
-
+	parent.Epilogue = epilogue
 	// If a Part is "multipart/" Content-Type, it will have .0 appended to its PartID
-	// i.e. it is the root of its MIME Part subtree
+	// i.e. it is the root of its MIME Part subtree.
 	if !firstRecursion {
 		parent.PartID += ".0"
 	}
-
 	return nil
 }